Buyingstore persistency Beta
Changed the loading process to clean up if there are some leftovers(corrupt tables).

Modified the loading query to also take in the zeny limit as criteria, since it would not make sense to set up an autotrader that already has reached his limit(this case should not be possible, if you do not modify the values in your MySQL table by hand).

Enjoy and do not forget to run the upgrade .sql files.