Pincode Functionality Extension
Added some options to allow or disallow certain combinations in the pincode as aegis has it.
In AEGIS it is not allowed to use pincodes like 0000 or 1111 where the pincode only contains 1 character. It is also not allowed to use sequences like 1234 or 5678.

In addition to the official check, which only disallows ascending sequences I also added descending ones.