Frost Joke pre-renewal freeze duration (follow-up to a4ee654) * The freeze duration of Frost Joke is now 12s in pre-renewal
Note: Each status change has a default duration on official servers and most skills use it. The standard duration for Freeze in pre-renewal is 12s, but it was increased to 30s in renewal. Both Storm Gust and Frost Joke could be proven to use the standard duration, whereas Frost Diver and Snowflake Draft have a custom duration per level set.